refactor: move e2e test fixture into common crate
